+++ frontend/src/App.vue
@@ -152,11 +152,6 @@
         :open="archivePanelOpen"
         @close="archivePanelOpen = false"
       />
-      <MobileSessionDrawer
-        :open="store.mobileDrawerOpen"
-        @close="store.mobileDrawerOpen = false"
-        @open-archive="archivePanelOpen = true; store.mobileDrawerOpen = false"
-      />
       <ScheduledTasksPanel
         :visible="scheduledTasksVisible"
         :create-target-tab-id="scheduledTaskCreateTargetTabId"
@@ -177,7 +172,6 @@ import MobileControls from '@/components/MobileControls.vue'
 import AgentWorkspaceView from '@/components/AgentWorkspaceView.vue'
 import ChatSidebar from '@/components/ChatSidebar.vue'
 import ArchivedSessionsPanel from '@/components/ArchivedSessionsPanel.vue'
-import MobileSessionDrawer from '@/components/MobileSessionDrawer.vue'
 import NetworkAccessMenu from '@/components/NetworkAccessMenu.vue'
 import ScheduledTasksPanel from '@/components/ScheduledTasksPanel.vue'
 import LoginView from '@/views/LoginView.vue'